+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# (unreleased)
+- Disable incorrectly implemented `textDocument/implementation` provider [csharp-language-server#4][https://github.com/razzmatazz/csharp-language-server/issues/4];
+
 # 0.1.6
 - Update roslyn+msbuild libs to support for `net6.0` projects;
 - Use .NET 6.0 for runtime;